Added some of the blocks together
only 2 =p
